requirements.txt
numpy>=1.21.0
pyyaml>=6.0

opencv-python>=4.5.0
Pillow>=9.0.0

insightface>=0.7.3
onnxruntime>=1.16.0
onnxruntime-gpu>=1.16.0 ; platform_system != "Darwin"

scipy>=1.8.0

customtkinter>=5.2.0
packaging
tqdm>=4.64.0